Fixed link for RSEs in Germany
The link for RSE groups in Germany seemed to be missing but the UK Society link seemed to be there as a placeholder. Replaced this with the link to the [map on the deRSE site](https://de-rse.org/en/map.html) which while not specifically highlighting RSE groups may help people to find a local or nearby RSE contact.